re-writing state structure to be entirely informed from containing application
