+++ packages/fast-figma-plugin-msft/src/interface/plugin-ui.state.ts
@@ -1,0 +1,6 @@
+/**
+ * Define the react state object for the Plugin UI
+ */
+export interface PluginUIState {
+    activeNodeType: NodeType | null;
+}
