+++ packages/fast-figma-plugin-msft/src/interface/strings.ts
